+++ src/model/user.rs
@@ -15,7 +15,7 @@ pub struct PublicUser {
     pub followers: Option<HashMap<String, Option<Value>>>,
     pub href: String,
     pub id: String,
-    pub images: Option<Vec<Image>>,
+    pub images: Vec<Image>,
     #[serde(rename = "type")]
     pub _type: Type,
     pub uri: String,
@@ -36,4 +36,37 @@ pub struct PrivateUser {
     #[serde(rename = "type")]
     pub _type: Type,
     pub uri: String,
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    #[test]
+    fn test_public_user() {
+        let json_str = r#"
+{
+  "display_name": "Ronald Pompa",
+  "external_urls": {
+    "spotify": "https://open.spotify.com/user/wizzler"
+  },
+  "followers": {
+    "href": null,
+    "total": 4050
+  },
+  "href": "https://api.spotify.com/v1/users/wizzler",
+  "id": "wizzler",
+  "images": [
+    {
+      "height": null,
+      "url": "https://i.scdn.co/image/ab6775700000ee85b5d374d281b9e510eda15fdf",
+      "width": null
+    }
+  ],
+  "type": "user",
+  "uri": "spotify:user:wizzler"
+}
+"#;
+        let user: PublicUser = serde_json::from_str(&json_str).unwrap();
+        assert_eq!(user.id, "wizzler".to_string());
+    }
 }